feat(linter): implement use_consistent_new_builtin
