+++ crates/biome_configuration/src/linter/rules.rs
@@ -2696,12 +2696,12 @@ pub struct Nursery {
     #[doc = "Disallow the use of dependencies that aren't specified in the package.json."]
     #[serde(skip_serializing_if = "Option::is_none")]
     pub no_undeclared_dependencies: Option<RuleConfiguration<NoUndeclaredDependencies>>,
+    #[doc = "Enforce the use of new for all builtins, except String, Number, Boolean, Symbol and BigInt."]
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub use_consistent_new_builtin: Option<RuleConfiguration<UseConsistentNewBuiltin>>,
     #[doc = "Disallows package private imports."]
     #[serde(skip_serializing_if = "Option::is_none")]
     pub use_import_restrictions: Option<RuleConfiguration<UseImportRestrictions>>,
-    #[doc = "Enforce the use of new for all builtins, except String, Number, Boolean, Symbol and BigInt."]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub use_new_for_builtins: Option<RuleConfiguration<UseNewForBuiltins>>,
     #[doc = "Enforce the sorting of CSS utility classes."]
     #[serde(skip_serializing_if = "Option::is_none")]
     pub use_sorted_classes: Option<RuleConfiguration<UseSortedClasses>>,
@@ -2739,8 +2739,8 @@ impl Nursery {
         "noReactSpecificProps",
         "noRestrictedImports",
         "noUndeclaredDependencies",
+        "useConsistentNewBuiltin",
         "useImportRestrictions",
-        "useNewForBuiltins",
         "useSortedClasses",
     ];
     const RECOMMENDED_RULES: &'static [&'static str] = &[
@@ -2879,12 +2879,12 @@ impl Nursery {
                 index_set.insert(RuleFilter::Rule(Self::GROUP_NAME, Self::GROUP_RULES[15]));
             }
         }
-        if let Some(rule) = self.use_import_restrictions.as_ref() {
+        if let Some(rule) = self.use_consistent_new_builtin.as_ref() {
             if rule.is_enabled() {
                 index_set.insert(RuleFilter::Rule(Self::GROUP_NAME, Self::GROUP_RULES[16]));
             }
         }
-        if let Some(rule) = self.use_new_for_builtins.as_ref() {
+        if let Some(rule) = self.use_import_restrictions.as_ref() {
             if rule.is_enabled() {
                 index_set.insert(RuleFilter::Rule(Self::GROUP_NAME, Self::GROUP_RULES[17]));
             }
@@ -2978,12 +2978,12 @@ impl Nursery {
                 index_set.insert(RuleFilter::Rule(Self::GROUP_NAME, Self::GROUP_RULES[15]));
             }
         }
-        if let Some(rule) = self.use_import_restrictions.as_ref() {
+        if let Some(rule) = self.use_consistent_new_builtin.as_ref() {
             if rule.is_disabled() {
                 index_set.insert(RuleFilter::Rule(Self::GROUP_NAME, Self::GROUP_RULES[16]));
             }
         }
-        if let Some(rule) = self.use_new_for_builtins.as_ref() {
+        if let Some(rule) = self.use_import_restrictions.as_ref() {
             if rule.is_disabled() {
                 index_set.insert(RuleFilter::Rule(Self::GROUP_NAME, Self::GROUP_RULES[17]));
             }
@@ -3093,12 +3093,12 @@ impl Nursery {
                 .no_undeclared_dependencies
                 .as_ref()
                 .map(|conf| (conf.level(), conf.get_options())),
-            "useImportRestrictions" => self
-                .use_import_restrictions
+            "useConsistentNewBuiltin" => self
+                .use_consistent_new_builtin
                 .as_ref()
                 .map(|conf| (conf.level(), conf.get_options())),
-            "useNewForBuiltins" => self
-                .use_new_for_builtins
+            "useImportRestrictions" => self
+                .use_import_restrictions
                 .as_ref()
                 .map(|conf| (conf.level(), conf.get_options())),
             "useSortedClasses" => self
